Modified section edit such that it works with Greebo
Section edit was re-factored in the Dokuwiki release “Greebo”, see https://www.dokuwiki.org/devel:section_editor. After updating Dokuwiki to "Greebo", the error "startSectionEdit: $data "section" is NOT an array!" appeared on top of each page that contains tabboxes. I modified the syntax part of the tabbox plugin such that it now works with "Greebo" and is still backward compatible.